[⚙️ chore] 페이지별 고정텍스트들 상수로 분리 #17
