+++ src/constants/challengeSectionText.ts
@@ -1,0 +1,16 @@
+const challengeSectionText = {
+  진행중: {
+    imogi: '🧑🏻‍💻 ',
+    subText: '현재 진행하고 있는 챌린지예요!',
+  },
+  대기중: {
+    imogi: '📚 ',
+    subText: '신청을 완료하고 시작을 대기중인 챌린지예요!',
+  },
+  진행완료: {
+    imogi: '🥳 ',
+    subText: '진행을 완료한 챌린지예요!',
+  },
+};
+
+export default challengeSectionText;
